♻️ Refactor Searchbar
- from emotion to vanilla-extract
